# sovereign-pm/openapi.yaml
openapi: "3.0.3"
info:
  title: Sovereign PM API
  version: "1.0.0"
  description: |
    AI-native project management API for the Sovereign Platform.
    Manages Projects, Epics, Stories, and AgentRuns. Generates prd.json
    for Ralph consumption and triggers on-demand Ralph loop runs via K8s Jobs.
  license:
    name: Apache-2.0
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak OIDC JWT (RS256) or HS256 test token
  schemas:
...
    AgentRun:
      type: object
      required: [id, storyId, status, logs, startedAt, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        storyId:
          type: string
          format: uuid
        status:
          type: string
          enum: [running, pass, fail]
        logs:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AgentRunCreate:
      type: object
      required: [storyId]
      properties:
        storyId:
          type: string
          format: uuid
        logs:
          type: string
        startedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /agent-runs:
    get:
      summary: List all agent runs
      operationId: listAgentRuns
      tags: [AgentRuns]
      responses:
        "200":
          description: List of agent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRun'
        "401":
          description: Unauthorized
    post:
      summary: Create an agent run record
      operationId: createAgentRun
      tags: [AgentRuns]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRunCreate'
      responses:
        "201":
          description: AgentRun created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        "400":
          description: Validation error
        "401":
          description: Unauthorized
  /agent-runs/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get an agent run by ID
      operationId: getAgentRun
      tags: [AgentRuns]
      responses:
        "200":
          description: AgentRun
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRun'
        "404":
          description: Not found
    put:
      summary: Update an agent run (mark pass/fail, append logs)
      operationId: updateAgentRun
      tags: [AgentRuns]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [running, pass, fail]
                logs:
                  type: string
                completedAt:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Updated agent run
...
  /ralph/runs:
    get:
      summary: List all Ralph run history
      operationId: listRalphRuns
      tags: [Ralph]
      responses:
        "200":
          description: List of agent runs ordered by startedAt desc
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentRun'
        "401":
          description: Unauthorized